% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_equations.R
\name{add_biomass}
\alias{add_biomass}
\title{Add biomass.}
\usage{
add_biomass(data, species, site, dbh_unit = guess_dbh_unit(data$dbh),
  biomass_unit = "kg")
}
\arguments{
\item{data}{A dataframe as those created with \code{\link[=add_equations]{add_equations()}}.}

\item{species}{A ForestGEO-like species-dataframe.}

\item{site}{The name of the site. One of \code{allodb::sites_info$site}.}

\item{dbh_unit}{Character string giving the unit of dbh values, e.g. "mm".}

\item{biomass_unit}{Character string giving the output unit e.g. "kg".}
}
\value{
A dataframe with a single row by each value of \code{rowid}.
}
\description{
Add biomass.
}
\examples{
library(dplyr)

data <- bind_rows(
  fgeo.biomass::scbi_stem_tiny_tree,
  fgeo.biomass::scbi_stem_tiny_shrub
)

species <- fgeo.biomass::scbi_species

add_biomass(data, species, site = "scbi")

data \%>\%
  add_biomass(species, site = "scbi", biomass_unit = "Mg") \%>\%
  select(species, biomass, dbh, everything())
}
