% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_species.R
\name{add_species}
\alias{add_species}
\title{Add \code{species} to a ForestGEO-like dataframe of census data.}
\usage{
add_species(census, species, site)
}
\arguments{
\item{census}{A ForestGEO-like census-dataframe.}

\item{species}{A ForestGEO-like species-dataframe.}

\item{site}{The name of the site. One of \code{allodb::sites_info$site}.}
}
\value{
A dataframe.
}
\description{
Add \code{species} to a ForestGEO-like dataframe of census data.
}
\examples{
census <- fgeo.biomass::scbi_tree1
species <- fgeo.biomass::scbi_species
add_species(census, species, site = "scbi")
}
\concept{constructors}
