% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tropical_biomass.R
\name{add_tropical_biomass}
\alias{add_tropical_biomass}
\title{Add biomass using allometric equations for tropical trees.}
\usage{
add_tropical_biomass(data, species, region = "Pantropical",
  latitude = NULL, longitude = NULL,
  dbh_unit = guess_dbh_unit(data$dbh))
}
\arguments{
\item{data}{A ForestGEO-like census-dataframe.}

\item{species}{A ForestGEO-like species-dataframe.}

\item{region}{Area of your dataset to estimate tree height thanks to Weibull-H region-, continent-specific
and pantropical models proposed by Feldpausch et al. (2012). To be chosen between:
\itemize{
\item \code{Africa}: Africa
\item \code{CAfrica}: Central Africa
\item \code{EAfrica}: Eastern Africa
\item \code{WAfrica}: Western Africa
\item \code{SAmerica}: Southern America
\item \code{BrazilianShield}: Brazilian Shield
\item \code{ECAmazonia}: East-Central Amazonia
\item \code{GuianaShield}: Guiana Shield
\item \code{WAmazonia}: Western Amazonia
\item \code{SEAsia}: South-Eastern Asia
\item \code{NAustralia}: Northern Australia
\item \code{Pantropical}: Pantropical
}}

\item{latitude, longitude}{A number giving coordinates, e.g. \code{latitude = 9.004080}, \code{longitude = -79.525635}. It can also be a vector of such
numbers, with as many elements as the number of rows of \code{data}.}

\item{dbh_unit}{Character string giving the unit of dbh values, e.g. "mm"
(see \code{\link[=valid_units]{valid_units()}}).}
}
\value{
A modified version of the \code{data} dataframe, with additional columns giving
taxonomic, wood density (in g/cm^3), and \code{biomass} (in kg) information.
}
\description{
This function wraps a number of functions from the
\link[BIOMASS:BIOMASS-package]{BIOMASS} package, which you may see for more
options and details. It uses pantropical models from Chave et al. 2014 to
estimate the above-ground biomass of tropical trees.
}
\examples{
library(dplyr)
library(fgeo.biomass)

data <- fgeo.biomass::scbi_stem_tiny_tree
species <- fgeo.biomass::scbi_species

add_tropical_biomass(data, species, region = "pantropical")

# Not running to reduce build check-time
\dontrun{
data \%>\%
  add_tropical_biomass(species, latitude = -34, longitude = -58) \%>\%
  select(biomass, everything())
}
}
\references{
Chave et al. (2014) \emph{Improved allometric models to estimate the aboveground
biomass of tropical trees}, Global Change Biology, 20 (10), 3177-3190
}
\seealso{
\code{\link[BIOMASS:computeAGB]{BIOMASS::computeAGB()}}, \code{\link[BIOMASS:retrieveH]{BIOMASS::retrieveH()}} \code{\link[=add_wood_density]{add_wood_density()}}.
}
