% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wood_density.R
\name{add_wood_density}
\alias{add_wood_density}
\title{Add wood density.}
\usage{
add_wood_density(data, species, family = TRUE, region = "World")
}
\arguments{
\item{data}{A ForestGEO-like census-dataframe.}

\item{species}{A ForestGEO-like species-dataframe.}

\item{family}{(optional) Vector of families. If set, the missing wood densities at the genus
level will be attributed at family level if available.}

\item{region}{Region (or vector of region) of interest of your sample. By default, Region is
set to 'World', but you can restrict the WD estimates to a single region :
\itemize{
\item \code{AfricaExtraTrop}: Africa (extra tropical)
\item \code{AfricaTrop}: Africa (tropical)
\item \code{Australia}: Australia
\item \code{AustraliaTrop}: Australia (tropical)
\item \code{CentralAmericaTrop}: Central America (tropical)
\item \code{China}: China
\item \code{Europe}: Europe
\item \code{India}: India
\item \code{Madagascar}: Madagascar
\item \code{Mexico}: Mexico
\item \code{NorthAmerica}: North America
\item \code{Oceania}: Oceania
\item \code{SouthEastAsia}: South-East Asia
\item \code{SouthEastAsiaTrop}: South-East Asia (tropical)
\item \code{SouthAmericaExtraTrop}: South America (extra tropical)
\item \code{SouthAmericaTrop}: South America (tropical)
\item \code{World}: World
}}
}
\value{
A dataframe as that passed to \code{data}, but with additional columns giving
taxonomic information, and the following wood density information (oven dry
mass/fresh volume in g/cm^3):
\itemize{
\item wd_mean: Mean wood density.
\item wd_sd: Standard deviation of the wood density that can be used in error
propagation (see sd_10 and AGBmonteCarlo()).
\item wd_level: Level at which wood density has been calculated. Can be
species, genus, family, dataset (mean of the entire dataset) or, if stand
is set, the name of the stand (mean of the current stand).
}
}
\description{
This function wraps some features of \code{\link[BIOMASS:getWoodDensity]{BIOMASS::getWoodDensity()}}, which you
may see for more details and options. It estimates the wood density of the
trees from their taxonomy or from their congeners using the global wood
density database (Chave et al. 2009, Zanne et al. 2009) or any additional
dataset. The resulting values of wood density can either be attributed to an
individual at the level of a species, genus, or family.
}
\details{
This function assigns to each taxon a species- or genus- level average if at
least one wood density value at the genus level is available for that taxon
in the reference database. If not, the mean wood density of the family (if
\code{family = TRUE}).

As an estimate of the error associated with the wood density estimate, this
function also provides the mean standard deviation value at the appropriate
taxonomic level.
}
\examples{
tree <- fgeo.biomass::scbi_stem_tiny_tree
species <- fgeo.biomass::scbi_species
add_wood_density(tree, species)
}
\references{
Rejou-Mechain M, Tanguy A, Piponiot C, Chave J, Herault B (2017). “BIOMASS :
an R package for estimating above-ground biomass and its uncertainty in
tropical forests.” \emph{Methods in Ecology and Evolution}, \emph{8}(9). ISSN 2041210X,
doi: 10.1111/2041-210X.12753 (URL: http://doi.org/10.1111/2041-210X.12753),
<URL: http://doi.wiley.com/10.1111/2041-210X.12753>.
}
\seealso{
BIOMASS::getWoodDensity
}
