% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allo_evaluate.R
\name{allo_evaluate}
\alias{allo_evaluate}
\title{Evaluate equations, giving a biomass result per row.}
\usage{
allo_evaluate(data, output_units = "kg")
}
\arguments{
\item{data}{A dataframe as those created with \code{\link[=allo_find]{allo_find()}}.}

\item{output_units}{Character string giving the output unit e.g. "kg".}
}
\value{
A dataframe with a single row by each value of \code{rowid}.
}
\description{
Evaluate equations, giving a biomass result per row.
}
\examples{
library(dplyr)

best <- fgeo.biomass::scbi_tree1 \%>\%
  add_species(fgeo.biomass::scbi_species, "scbi") \%>\%
  allo_find()

allo_evaluate(best)
}
\seealso{
Other functions to manipulate equations: \code{\link{add_equations}},
  \code{\link{allo_find}}, \code{\link{allo_lookup}}
}
\concept{functions to manipulate equations}
