% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allo_find.R
\name{allo_find}
\alias{allo_find}
\title{Get default equations of each type.}
\usage{
allo_find(dbh_species, custom_eqn = NULL)
}
\arguments{
\item{dbh_species}{A dataframe as those created with \code{\link[=add_species]{add_species()}}.}

\item{custom_eqn}{A dataframe of class "eqn".}
}
\value{
A nested dataframe with each row containing the data of an equation
type.
}
\description{
Get default equations of each type.
}
\examples{
census <- fgeo.biomass::scbi_tree1
species <- fgeo.biomass::scbi_species
dbh_species <- add_species(
  census, species,
  site = "scbi"
)

allo_find(dbh_species)
}
\seealso{
Other functions to manipulate equations: \code{\link{add_equations}},
  \code{\link{allo_evaluate}}, \code{\link{allo_lookup}},
  \code{\link{allo_order}}

Other constructors: \code{\link{add_species}}
}
\concept{constructors}
\concept{functions to manipulate equations}
