% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allo_lookup.R
\name{allo_lookup}
\alias{allo_lookup}
\title{Look up in \strong{allodb} information about each \code{equation_id}.}
\usage{
allo_lookup(data, allodb = allodb::equations)
}
\arguments{
\item{data}{A dataframe with columns \code{rowid} and \code{equation_id}}

\item{allodb}{A table from allodb.}
}
\value{
A dataframe.
}
\description{
Look up in \strong{allodb} information about each \code{equation_id}.
}
\examples{
library(dplyr)

census <- fgeo.biomass::scbi_tree1 \%>\% sample_n(500)
species <- fgeo.biomass::scbi_species

best <- census \%>\%
  add_species(species, site = "scbi") \%>\%
  allo_find() \%>\%
  fixme_drop_duplicated_rowid()

ids <- add_equations(census, best)
allo_lookup(ids)
}
\seealso{
Other functions to manipulate equations: \code{\link{add_equations}},
  \code{\link{allo_evaluate}}, \code{\link{allo_find}}
}
\concept{functions to manipulate equations}
