% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allo_order.R
\name{allo_order}
\alias{allo_order}
\title{Find the best equations given a priority order.}
\usage{
allo_order(.data, order = NULL)
}
\arguments{
\item{.data}{List of dataframes. Each should have a \code{rowid} column giving the
index of each row. Otherwise, \code{rowid} will be added with a warning.}

\item{order}{String giving the name or index of the list elements in the
order they should be row-bind, with elements on the left winning over
elements on the right.}
}
\value{
A dataframe.
}
\description{
This function orders the output of \code{\link[=allo_find]{allo_find()}} by equation type, then
reduces it in order, by overwriting equations of each type with the
equations of a higher-priority type.
}
\examples{
allodb::scbi_tree1 \%>\%
  add_species(allodb::scbi_species, "scbi") \%>\%
  allo_find() \%>\%
  allo_order()
}
\seealso{
\code{\link[=reduce_complement]{reduce_complement()}}.

Other functions to manipulate equations: \code{\link{add_equations}},
  \code{\link{allo_evaluate}}, \code{\link{allo_find}},
  \code{\link{allo_lookup}}
}
\concept{functions to manipulate equations}
