% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_species.R
\name{allodb_eqn_crucial}
\alias{allodb_eqn_crucial}
\title{Crucial columns form \strong{allodb} equations-table.}
\usage{
allodb_eqn_crucial()
}
\value{
A string.
}
\description{
Crucial columns form \strong{allodb} equations-table.
}
\examples{
allodb_eqn_crucial()
}
\keyword{internal}
