% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad_eqn_id.R
\name{bad_eqn_id}
\alias{bad_eqn_id}
\title{Find equations that can't be evaluated.}
\usage{
bad_eqn_id(.data)
}
\arguments{
\item{.data}{An \strong{allodb} equations-table (e.g. allodb::master()).}
}
\value{
A character vector.
}
\description{
Find equations that can't be evaluated.
}
\examples{
bad_eqn_id(allodb::master())
}
\seealso{
Other internal objects that will be hidden or removed: \code{\link{.bad_eqn_id}},
  \code{\link{.default_eqn}}, \code{\link{default_eqn}},
  \code{\link{handle_multiple_rowid}}
}
\concept{internal objects that will be hidden or removed}
