% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_species.R
\name{census_species}
\alias{census_species}
\title{Add \code{species} to a ForestGEO-like dataframe of census data.}
\usage{
census_species(census, species, site)
}
\arguments{
\item{census}{A ForestGEO-like census-dataframe.}

\item{species}{A ForestGEO-like species-dataframe.}

\item{site}{The name of the site. One of \code{allodb::sites_info$site}.}
}
\value{
A dataframe.
}
\description{
Add \code{species} to a ForestGEO-like dataframe of census data.
}
\examples{
census_species(allodb::scbi_tree1, allodb::scbi_species, site = "scbi")
}
\concept{constructors}
