% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement}
\alias{complement}
\title{Complement a dataframe with rows from another dataframe.}
\usage{
complement(data, complement)
}
\arguments{
\item{data}{Dataframes which rows to always keep.}

\item{complement}{Dataframe which rows will be used to complement \code{data}.}
}
\value{
A dataframe.
}
\description{
Complement a dataframe with rows from another, identically structured
dataframe. This is the result of reducing multiple joins:
\itemize{
\item \code{dplyr::semi_join(data, complement)}.
\item \code{dplyr::anti_join(data, complement)}.
\item \code{dplyr::anti_join(complement, data)}.
Each dataframe must have a \code{rowid}
}
}
\examples{
library(dplyr)

dfm <- tibble(rowid = 1, from = "dfm")
cmp <- tibble(rowid = 1:2, from = "cmp")
dfm \%>\% complement(cmp)

dfm2 <- tibble(rowid = 1:2, x = "dfm2")
cmp2 <- tibble(rowid = 1:2, x = "cmp2")
complement(dfm2, cmp2)

traffic_light <- tribble(
~rowid, ~color,
     1, "red",
     2, "yellow"
)

more_colors <- tribble(
~rowid, ~color,
     1, "blue",
     2, "orange",
     3, "green"
)

traffic_light \%>\% complement(more_colors)
}
\seealso{
\code{\link[dplyr:join]{dplyr::join()}}, \code{\link[tibble:rowid_to_column]{tibble::rowid_to_column()}}.

Other helpers: \code{\link{reduce_complement}}
}
\concept{helpers}
