% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{A safe version of \code{measurements::conv_unit()}.}
\usage{
convert_units(x, from, to)
}
\arguments{
\item{x}{a numeric vector giving the measurement value in its original units.}

\item{from}{the unit in which the measurement was made.}

\item{to}{the unit to which the measurement is to be converted.}
}
\value{
As in \code{measurements::conv_unit()} if possible, or NA.
}
\description{
Unlike \code{measurements::conv_unit()} this function doesn't fail if it can't
convert a unit -- instead it outputs NA.
}
\examples{
convert_units(c(1, NA), from = "cm", to = "m")
# Same
measurements::conv_unit(c(1, NA), from = "cm", to = "m")

# Passes
convert_units(c(1, 10), from = "cm", to = c("m", "bad"))
# Errs
try(measurements::conv_unit(c(1, 10), from = "cm", to = c("m", "bad")))
}
