% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_eqn.R
\name{crucial_equation_cols}
\alias{crucial_equation_cols}
\title{Crucial columns from \strong{allodb} equations-table.}
\usage{
crucial_equation_cols()
}
\value{
A string.
}
\description{
Crucial columns from \strong{allodb} equations-table.
}
\examples{
crucial_equation_cols()
}
\keyword{internal}
