% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_eqn.R
\name{default_eqn}
\alias{default_eqn}
\title{Restructure equations from \strong{allodb}.}
\usage{
default_eqn(data)
}
\arguments{
\item{data}{\code{\link[allodb:master_tidy]{allodb::master_tidy()}} or similar.}
}
\value{
A dataframe.
}
\description{
This function restructures an equations-table from \strong{allodb} with columns as
in \code{\link[=allodb_eqn_crucial]{allodb_eqn_crucial()}} (e.g. \code{\link[allodb:master_tidy_tidy]{allodb::master_tidy_tidy()}}). It transforms its
input into a default-equations table. Now this function is very strict and
intrusive:
\itemize{
\item It drops problematic equations that can't be evaluated.
\item It adds and remove columns.
\item It renames columns.
\item It transforms text-values to lowercase to simplify matching.
\item It re-formats the text-representation of equations.
\item It drops missing values.
\item It replaces spaces (" ") with underscore ("_") in values of
allometry_specificity for easier manipulation.
}
}
\examples{
default_eqn(allodb::master_tidy())
}
\concept{internal objects that will be hidden or removed}
