% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_equations.R
\docType{data}
\name{default_equations}
\alias{default_equations}
\title{For speed, save store the result of \code{default_eqn(allodb::master())}.}
\format{An object of class \code{eqn} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 614 rows and 9 columns.}
\usage{
default_equations
}
\description{
For speed, save store the result of \code{default_eqn(allodb::master())}.
}
\keyword{datasets}
