% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\name{.bad_eqn_id}
\alias{.bad_eqn_id}
\title{Character vector giving \code{equation_id} of equations that can't be evaluated.}
\description{
Output of \code{\link[=bad_eqn_id]{bad_eqn_id()}} -- stored for speed.
}
\seealso{
Other internal objects that will be hidden or removed: \code{\link{.default_eqn}},
  \code{\link{bad_eqn_id}}, \code{\link{default_eqn}},
  \code{\link{handle_multiple_rowid}}
}
\concept{internal objects that will be hidden or removed}
