% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.biomass-package.R
\docType{package}
\name{fgeo.biomass-package}
\alias{fgeo.biomass}
\alias{fgeo.biomass-package}
\title{fgeo.biomass: Calculate Biomass with the Best Available Allometries}
\description{
Calculate biomass with the best available
    allometries.  allodb package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.biomass}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.biomass/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} (0000-0002-1986-7988) [contractor]

Authors:
\itemize{
  \item Gabriel Arellano \email{gabriel.arellano.torres@gmail.com} [reviewer]
  \item Stuart Davies \email{daviess@si.edu} [reviewer]
  \item Erika Gonzalez \email{GonzalezEB@si.edu} (0000-0001-8305-6672) [reviewer]
  \item Sean McMahon \email{mcmahons@si.edu} [reviewer]
  \item Kristina Anderson-Teixeira \email{TeixeiraK@si.edu} [reviewer]
}

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
