% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_units.R
\name{fix_units}
\alias{fix_units}
\title{Try to convert units to conform with \code{measurements::conv_unit_options}.}
\usage{
fix_units(x)
}
\arguments{
\item{x}{A character vector of units.}
}
\value{
A character vector.
}
\description{
\code{fix_units()} leaves valid units untouched, fix invalid known units, and
flags invalid unknown units.
}
\examples{
x <- c("mm", "cm", "m", "in", "in^2", "cm^2", "BAD")
fix_units(x)

unique(
  allodb::equations$dbh_units_original
)

unique(
  fix_units(allodb::equations$dbh_units_original)
)

unique(
  fix_units(allodb::equations$biomass_units_original)
)
}
\seealso{
\code{\link[=valid_units]{valid_units()}}.
}
\keyword{internal}
