% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_equations.R
\name{fixme_exclude_failing_equations}
\alias{fixme_exclude_failing_equations}
\title{Exclude equations that can't be evaluated.}
\usage{
fixme_exclude_failing_equations(data)
}
\arguments{
\item{data}{An equations dataframe, e.g. \code{allodb::master()}.}
}
\value{
A dataframe with all equations that can be successfully evaluated.
}
\description{
This function tries to evaluate \strong{allodb} equations and excludes the ones
that err.
}
\examples{
fixme_exclude_failing_equations(allodb::master())
}
\seealso{
Other internal functions that flag issues to be fixed: \code{\link{fixme_pull_failing_eqn}},
  \code{\link{handle_multiple_rowid}}
}
\concept{internal functions that flag issues to be fixed}
