% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixme_pull_failing_eqn.R
\name{fixme_pull_failing_eqn}
\alias{fixme_pull_failing_eqn}
\title{Find failing equations, that can't be evaluated.}
\usage{
fixme_pull_failing_eqn(data)
}
\arguments{
\item{data}{An \strong{allodb} equations-table (e.g. allodb::master()).}
}
\value{
A character vector.
}
\description{
Find failing equations, that can't be evaluated.
}
\examples{
fixme_pull_failing_eqn(allodb::master())
}
\seealso{
Other internal functions that flag issues to be fixed: \code{\link{handle_multiple_rowid}}
}
\concept{internal functions that flag issues to be fixed}
