% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixme_units.R
\name{fixme_units}
\alias{fixme_units}
\title{Try to convert units to conform with \code{measurements::conv_unit_options}.}
\usage{
fixme_units(x)
}
\arguments{
\item{x}{A character vector of units.}
}
\value{
A character vector.
}
\description{
\code{fixme_units()} leaves valid units untouched, fix invalid known units, and
flags invalid unknown units.
}
\examples{
x <- c("mm", "cm", "m", "in", "in^2", "cm^2", "BAD")
fixme_units(x)

unique(
  allodb::equations$dbh_units_original
)

unique(
  fixme_units(allodb::equations$dbh_units_original)
)

unique(
  fixme_units(allodb::equations$biomass_units_original)
)
}
\seealso{
\code{\link[=valid_units]{valid_units()}}.

Other internal functions that flag issues to be fixed: \code{\link{fixme_pull_failing_eqn}}
}
\concept{internal functions that flag issues to be fixed}
