% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equations.R
\name{get_equations}
\alias{get_equations}
\title{Get default equations of each type.}
\usage{
get_equations(dbh_species)
}
\arguments{
\item{dbh_species}{A dataframe as those created with \code{\link[=census_species]{census_species()}}.}
}
\value{
A nested dataframe with each row containing the data of an equation
type.
}
\description{
Get default equations of each type.
}
\examples{
census <- allodb::scbi_tree1
species <- allodb::scbi_species
dbh_species <- census_species(
  census, species,
  site = "scbi"
)

get_equations(dbh_species)
}
\seealso{
Other functions to manipulate equations: \code{\link{add_equations}},
  \code{\link{evaluate_equations}},
  \code{\link{pick_best_equations}}
}
\concept{functions to manipulate equations}
