% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_dbh_unit.R
\name{guess_dbh_unit}
\alias{guess_dbh_unit}
\title{Guess \code{dbh} units based on minimum and maximum values of \code{dbh}.}
\usage{
guess_dbh_unit(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A length-1 character string.
}
\description{
Guess \code{dbh} units based on minimum and maximum values of \code{dbh}.
}
\examples{
# Case when
# min(x, na.rm = TRUE) < 1.1 &&
# max(x, na.rm = TRUE) < 500
guess_dbh_unit(c(1.0, 100))

# min is too large
try(guess_dbh_unit(c(1.2, 499)))
# max is too large
try(guess_dbh_unit(c(1.0, 501)))
# min is too large and max is too large
try(guess_dbh_unit(c(1.2, 501)))



# Case when
# min(x, na.rm = TRUE) > 9 &&
# max(x, na.rm = TRUE) > 500
guess_dbh_unit(c(9.1, 500.1))

# min is too small
try(guess_dbh_unit(c(8.9, 500.1)))
# max is too small
try(guess_dbh_unit(c(9.1, 500)))
# min is too small and max is too small
try(guess_dbh_unit(c(8.9, 500)))
}
