% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicated_rowid.R
\name{handle_multiple_rowid}
\alias{handle_multiple_rowid}
\alias{find_duplicated_rowid}
\alias{pick_one_row_by_rowid}
\title{Handle multiple \code{rowid}s: find duplicates and pick a single row.}
\usage{
find_duplicated_rowid(.data)

pick_one_row_by_rowid(.data)
}
\arguments{
\item{.data}{A dataframe as those created with \code{\link[=pick_best_equations]{pick_best_equations()}}.}
}
\value{
A dataframe with a single row by each value of \code{rowid}.
}
\description{
Handle multiple \code{rowid}s: find duplicates and pick a single row.
}
\examples{
library(dplyr)

best <- allodb::scbi_tree1 \%>\%
  add_species(allodb::scbi_species, "scbi") \%>\%
  allo_find() \%>\%
  pick_best_equations()

evaluate_equations(best)
}
\seealso{
Other internal objects that will be hidden or removed: \code{\link{.bad_eqn_id}},
  \code{\link{.default_eqn}}, \code{\link{bad_eqn_id}},
  \code{\link{default_eqn}}
}
\concept{internal objects that will be hidden or removed}
