% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_height.R
\name{model_height}
\alias{model_height}
\title{Fit a height-diameter model.}
\usage{
model_height(data, method = c("log1", "log2", "weibull", "michaelis"))
}
\arguments{
\item{data}{A ForestGEO-like census-dataframe.}

\item{method}{A character string. One of these:
\itemize{
\item \code{log 1} (equivalent to a power model): log(H) = a + b * log(D)
\item \code{log 2}: log(H) = a + b * log(D) + c * log(D)^2
\item \code{weibull}: H = a * (1 - exp(-(D / b) ^ c))
\item \code{michaelis}: H = (A * D) / (B + D)
}}
}
\value{
A list with these elements:
\itemize{
\item \code{input}: list of the data used to construct the model (list(H, D)).
\item \code{model}: outputs of the model (same outputs as given by stats::lm(),
stats::nls()).
\item \code{RSE}: Residual Standard Error of the model.
\item \code{RSElog}: Residual Standard Error of the log model (NULL if other model).
\item \code{residuals}: Residuals of the model.
\item \code{coefficients}: Coefficients of the model.
\item \code{R.squared}: R^2 of the model.
\item \code{formula}: Formula of the model.
\item \code{method}: Name of the method used to construct the model.
\item \code{predicted}: Predicted height values.
}
}
\description{
This functions wraps \code{\link[BIOMASS:modelHD]{BIOMASS::modelHD()}}, which you can see for more
options and details.
}
\details{
All the back transformations for log-log models are done using the Baskerville correction (\eqn{0.5 * RSE^2},
where RSE is the Residual Standard Error).
}
\examples{
data <- fgeo.biomass::scbi_tree1[1:30, ]

str(model_height(data))

out <- model_height(data, method = "log2")
# Avoid too much output
str(out[-2])
# Glimpse what wasn't shown above
lapply(out[2], names)
}
\seealso{
\code{\link[BIOMASS:modelHD]{BIOMASS::modelHD()}}.
}
