% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate_errors.R
\name{propagate_errors}
\alias{propagate_errors}
\title{Propagate errors in above-ground biomass (AGB).}
\usage{
propagate_errors(data, n = 1000, dbh_sd = NULL, height_model = NULL)
}
\arguments{
\item{data}{The output of \code{\link[=add_tropical_biomass]{add_tropical_biomass()}}.}

\item{n}{Number of iterations. Cannot be smaller than 50 or larger than 1000. By default \code{n = 1000}}

\item{dbh_sd}{This variable can take three kind of values, indicating how to
propagate the errors on diameter measurements: a single numerical value or
a vector of the same size as D, both representing the standard deviation
associated with the diameter measurements or "chave2004" (an important
error on 5 percent of the measures, a smaller error on 95 percent of the
trees).}

\item{height_model}{Model used to estimate tree height from tree diameter
(output from \code{\link[=model_height]{model_height()}}, see example).}
}
\value{
A list with the following elements:
\itemize{
\item \code{meanAGB}: Mean stand AGB value following the error propagation.
\item \code{medAGB}: Median stand AGB value following the error propagation.
\item \code{sdAGB}: Standard deviation of the stand AGB value following the error
propagation.
\item \code{credibilityAGB}: Credibility interval at 95% of the stand AGB value
following the error propagation.
\item \code{AGB_simu}: Matrix with the AGB of the trees (rows) times the n iterations
(columns).
}
}
\description{
This functions is warps \code{\link[BIOMASS:AGBmonteCarlo]{BIOMASS::AGBmonteCarlo()}}, which you may use
directly for more options.
}
\details{
See Rejou-Mechain et al. (2017) for all details on the error propagation procedure.
}
\examples{
library(dplyr)

data <- fgeo.biomass::scbi_tree1 \%>\%
  slice(1:100)
species <- fgeo.biomass::scbi_species

# Using `region` (default)
biomass <- add_tropical_biomass(data, species)
model <- model_height(biomass)
str(
  propagate_errors(biomass, n = 50, height_model = model)
)

# Using `latitude` and `longitude`
biomass <- add_tropical_biomass(
  data = data,
  species = species,
  latitude = 4,
  longitude = -52
)

model <- model_height(biomass)

# Asks to confirm using the model instead of coordinates
if (interactive()) {
  str(
    propagate_errors(biomass, n = 50, height_model = model)
  )
}
}
\references{
Chave, J. et al. (2004). \emph{Error propagation and scaling for tropical forest biomass estimates}.
Philosophical Transactions of the Royal Society B: Biological Sciences, 359(1443), 409-420.

Rejou-Mechain et al. (2017).
\emph{BIOMASS: An R Package for estimating above-ground biomass and its uncertainty in tropical forests}.
Methods in Ecology and Evolution, 8 (9), 1163-1167.
}
