% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate_errors.R
\name{propagete_errors}
\alias{propagete_errors}
\title{Title}
\usage{
propagete_errors(data, n = 1000, dbh_sd = NULL, height_model = NULL)
}
\arguments{
\item{data}{The output of \code{\link[=add_tropical_biomass]{add_tropical_biomass()}}.}

\item{n}{Number of iterations. Cannot be smaller than 50 or larger than 1000. By default \code{n = 1000}}

\item{dbh_sd}{This variable can take three kind of values, indicating how to
propagate the errors on diameter measurements: a single numerical value or
a vector of the same size as D, both representing the standard deviation
associated with the diameter measurements or "chave2004" (an important
error on 5 percent of the measures, a smaller error on 95 percent of the
trees).}

\item{height_model}{}
}
\value{
A list with the following components:
\itemize{
\item \code{meanAGB}: Mean stand AGB value following the error propagation
\item \code{medAGB}: Median stand AGB value following the error propagation
\item \code{sdAGB}: Standard deviation of the stand AGB value following the error propagation
\item \code{credibilityAGB}: Credibility interval at 95% of the stand AGB value following the error propagation
\item \code{AGB_simu}: Matrix with the AGB of the trees (rows) times the n iterations (columns)
}
}
\description{
Title
}
\examples{
library(dplyr)

data <- fgeo.biomass::scbi_tree1 \%>\%
  slice(1:100)
species <- fgeo.biomass::scbi_species

biomass <- add_tropical_biomass(
  data = data,
  species = species,
  latitude = 4,
  longitude = -52
)

out <- propagete_errors(biomass, n = 50, height_model = NULL)
str(out)
}
