% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad_eqn_id.R
\name{pull_failing_eqn}
\alias{pull_failing_eqn}
\title{Find equations that can't be evaluated.}
\usage{
pull_failing_eqn(data)
}
\arguments{
\item{data}{An \strong{allodb} equations-table (e.g. allodb::master()).}
}
\value{
A character vector.
}
\description{
Find equations that can't be evaluated.
}
\examples{
pull_failing_eqn(allodb::master())
}
\seealso{
Other internal objects that will be hidden or removed: \code{\link{.default_eqn}},
  \code{\link{default_eqn}}, \code{\link{failing_eqn_id}}
}
\concept{internal objects that will be hidden or removed}
