% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_complement.R
\name{reduce_complement}
\alias{reduce_complement}
\title{Reduce a list of dataframes complementing each one with the following one.}
\usage{
reduce_complement(.x, order = NULL)
}
\arguments{
\item{.x}{List of dataframes. Each should have a \code{rowid} column giving the
index of each row, or \code{rowid} will be added with a warning.}

\item{order}{The order to use when reducing the list of dataframes given
either as a numeric vector of the index of list elements, or as a string
giving the names of the list elements. Missing elements are excluded.}
}
\value{
A dataframe.
}
\description{
Reduce a list of dataframes complementing each dataframe of the list with
rows from the next dataframe in the list. You can optionally set the \code{order}
to specify which elements to include, and in which order they will be
reduced.
}
\examples{
suppressPackageStartupMessages(
  library(dplyr)
)

game <- tibble::tribble(
  ~round, ~player,            ~result,
       1, "Ana",              "piedra",
       1, "Maria",            "papel",
       1, "Jose",             "tijera",

       2, "Maria; Jose",      "papel",
       2, "Ana",              "tijera",

       3, "Ana; Maria; Jose", "tijera"
)

# Who wins each round?
game \%>\%
  split(.$result) \%>\%
  reduce_complement()
}
\seealso{
Other helpers: \code{\link{complement}}
}
\concept{helpers}
