% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scbi}
\alias{scbi}
\alias{scbi_tree1}
\alias{scbi_species}
\alias{scbi_stem_shrub_tiny}
\alias{scbi_stem_tree_tiny}
\title{Data from SCBI.}
\format{Objects of class tbl_df (inherit from tbl, data.frame). See structure
in \strong{Examples}.}
\source{
http://bit.ly/scbi-forestgeo-data
}
\usage{
scbi_tree1

scbi_species

scbi_stem_shrub_tiny

scbi_stem_tree_tiny
}
\description{
\itemize{
\item \code{scbi_tree1} and \code{scbi_species} are complete tree and species tables from
SCBI.
\item \code{scbi_stem_shrub_tiny} and \code{scbi_stem_tree_tiny} are small example datasets
containing exclusively shrubs and trees.
}
}
\examples{
str(scbi_tree1)

str(scbi_stem_shrub_tiny)

str(scbi_stem_tree_tiny)

str(scbi_species)
}
\concept{census datasets}
\keyword{datasets}
