% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_units.R
\name{valid_units}
\alias{valid_units}
\title{Valid units as per \code{measurements::conv_unit_options}.}
\usage{
valid_units()
}
\value{
A list
}
\description{
This simply wraps \code{measurements::conv_unit_options} to provide a more
evocative name, closer to the problem domain. You may use this function to
know how to name units in a way that ensures that units conversion will be
converted correctly.
}
\examples{
valid_units()
}
\seealso{
Other helpers: \code{\link{convert_units}}
}
\concept{helpers}
