% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_equations.R
\name{add_equations}
\alias{add_equations}
\title{Add allometric equations to a census dataset.}
\usage{
add_equations(census, equations)
}
\arguments{
\item{census}{A ForestGEO-like census dataframe.}

\item{equations}{An equations dataframe with unique \code{rowid}s.}
}
\value{
A dataframe.
}
\description{
Add allometric equations to a census dataset.
}
\examples{
census <- allodb::scbi_tree1
species <- allodb::scbi_species

single_best <- census \%>\%
  census_species(species, site = "scbi") \%>\%
  get_equations() \%>\%
  pick_best_equations() \%>\%
  pick_one_row_by_rowid()

add_equations(census, single_best)
}
\seealso{
Other functions to manipulate equations: \code{\link{evaluate_equations}},
  \code{\link{get_equations}},
  \code{\link{pick_best_equations}}
}
\concept{functions to manipulate equations}
