% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_equations.R
\name{add_equations}
\alias{add_equations}
\title{Find allometric equations in allodb or in a custom equations-table.}
\usage{
add_equations(data, dbh_unit = NULL)
}
\arguments{
\item{data}{A dataframe as those created with \code{\link[=add_species]{add_species()}}.}

\item{dbh_unit}{Character string giving the unit of dbh values, e.g. "mm".}
}
\value{
A nested dataframe with each row containing the data of an equation
type.
}
\description{
Find allometric equations in allodb or in a custom equations-table.
}
\examples{
census <- dplyr::sample_n(fgeo.biomass::scbi_tree1, 30)
species <- fgeo.biomass::scbi_species
census_species <- add_species(
  census, species,
  site = "scbi"
)

add_equations(census_species, dbh_unit = "mm")
}
\seealso{
Other constructors: \code{\link{add_species}}
}
\concept{constructors}
\concept{functions to manipulate equations}
