% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allo_evaluate.R
\name{allo_evaluate}
\alias{allo_evaluate}
\title{Evaluate equations, giving a biomass result per row.}
\usage{
allo_evaluate(data, dbh_unit = "mm", biomass_unit = "kg")
}
\arguments{
\item{data}{A dataframe as those created with \code{\link[=allo_find]{allo_find()}}.}

\item{dbh_unit}{Character string giving the unit of the expected input dbh.}

\item{biomass_unit}{Character string giving the output unit e.g. "kg".}
}
\value{
A dataframe with a single row by each value of \code{rowid}.
}
\description{
Evaluate equations, giving a biomass result per row.
}
\examples{
library(dplyr)

best <- fgeo.biomass::scbi_tree1 \%>\%
  # Pick few rows for a quick example
  sample_n(500) \%>\%
  add_species(fgeo.biomass::scbi_species, "scbi") \%>\%
  allo_find()

allo_evaluate(best)

allo_evaluate(best, dbh_unit = "cm", biomass_unit = "Mg")
}
\seealso{
Other functions to manipulate equations: \code{\link{allo_find}}
}
\concept{functions to manipulate equations}
