% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\name{.default_eqn}
\alias{.default_eqn}
\title{Dataframe of default equations, restructured from \strong{allodb}.}
\description{
Output of \code{\link[=default_eqn]{default_eqn()}} -- stored for speed.
}
\seealso{
Other internal objects that will be hidden or removed: \code{\link{.bad_eqn_id}},
  \code{\link{bad_eqn_id}}, \code{\link{default_eqn}},
  \code{\link{handle_multiple_rowid}}
}
\concept{internal objects that will be hidden or removed}
