% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_equations.R
\name{evaluate_equations}
\alias{evaluate_equations}
\title{Evaluate equations, giving a biomass result per row.}
\usage{
evaluate_equations(.data)
}
\arguments{
\item{.data}{A dataframe as those created with \code{\link[=pick_best_equations]{pick_best_equations()}}.}
}
\value{
A dataframe with a single row by each value of \code{rowid}.
}
\description{
Evaluate equations, giving a biomass result per row.
}
\examples{
library(dplyr)

best <- allodb::scbi_tree1 \%>\%
  add_species(allodb::scbi_species, "scbi") \%>\%
  allo_find() \%>\%
  pick_best_equations()

evaluate_equations(best)
}
\seealso{
Other functions to manipulate equations: \code{\link{add_equations}},
  \code{\link{allo_find}}, \code{\link{allo_lookup}},
  \code{\link{pick_best_equations}}
}
\concept{functions to manipulate equations}
