% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixme_units.R
\name{fixme_units}
\alias{fixme_units}
\alias{valid_units}
\title{Try to convert units to conform with \code{measurements::conv_unit_options}.
\itemize{
\item \code{valid_units()} is an evocative alias of \code{measurements::conv_unit_options}.
\item \code{fixme_units()} leaves valid units untouched, fix invalid known units, and
flags invalid unknown units.
}}
\usage{
fixme_units(x)

valid_units()
}
\arguments{
\item{x}{A character vector of units.}
}
\value{
A character vector.
}
\description{
Try to convert units to conform with \code{measurements::conv_unit_options}.
\itemize{
\item \code{valid_units()} is an evocative alias of \code{measurements::conv_unit_options}.
\item \code{fixme_units()} leaves valid units untouched, fix invalid known units, and
flags invalid unknown units.
}
}
\examples{
x <- c("mm", "cm", "m", "in", "in^2", "cm^2", "BAD")
fixme_units(x)

unique(
  allodb::equations$dbh_units_original
)

unique(
  fixme_units(allodb::equations$dbh_units_original)
)

# All valid unites
valid_units()
}
\seealso{
Other functions to manipulate equations: \code{\link{add_equations}},
  \code{\link{allo_evaluate}}, \code{\link{allo_find}},
  \code{\link{allo_lookup}}
}
\concept{functions to manipulate equations}
