% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixme_find_duplicated_rowid.R
\name{handle_multiple_rowid}
\alias{handle_multiple_rowid}
\alias{fixme_find_duplicated_rowid}
\alias{fixme_pick_one_row_by_rowid}
\title{Handle multiple \code{rowid}s: find duplicates and pick a single row.}
\usage{
fixme_find_duplicated_rowid(.data)

fixme_pick_one_row_by_rowid(.data)
}
\arguments{
\item{.data}{A dataframe as those created with \code{\link[=allo_order]{allo_order()}}.}
}
\value{
A dataframe with a single row by each value of \code{rowid}.
}
\description{
Handle multiple \code{rowid}s: find duplicates and pick a single row.
}
\examples{
library(dplyr)

best <- fgeo.biomass::scbi_tree1 \%>\%
  add_species(fgeo.biomass::scbi_species, "scbi") \%>\%
  allo_find() \%>\%
  allo_order()

allo_evaluate(best)
}
\seealso{
Other functions to manipulate equations: \code{\link{add_equations}},
  \code{\link{allo_find}}, \code{\link{allo_lookup}},
  \code{\link{allo_order}}
}
\concept{internal functions that flag issues to be fixed}
