% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixme_find_duplicated_rowid.R
\name{handle_multiple_rowid}
\alias{handle_multiple_rowid}
\alias{fixme_find_duplicated_rowid}
\alias{fixme_drop_duplicated_rowid}
\title{Handle multiple \code{rowid}s: find duplicates and pick a single row.}
\usage{
fixme_find_duplicated_rowid(data)

fixme_drop_duplicated_rowid(data)
}
\arguments{
\item{data}{A dataframe as those created with \code{\link[=allo_find]{allo_find()}}.}
}
\value{
\itemize{
\item \code{fixme_find_duplicated_rowid()} returns a dataframe with the rows for
which duplicated \code{rowid} values are found.
\item \code{fixme_drop_duplicated_rowid()} returns a dataframe with a single row per
rowid value.
}
}
\description{
Handle multiple \code{rowid}s: find duplicates and pick a single row.
}
\examples{
best <- fgeo.biomass::scbi_tree1 \%>\%
  add_species(fgeo.biomass::scbi_species, "scbi") \%>\%
  allo_find()

best \%>\%
  fixme_find_duplicated_rowid()

best \%>\%
  fixme_drop_duplicated_rowid()

# Should return 0-rows
best \%>\%
  fixme_drop_duplicated_rowid() \%>\%
  fixme_find_duplicated_rowid()
}
\seealso{
Other internal functions that flag issues to be fixed: \code{\link{fixme_pull_failing_eqn}},
  \code{\link{fixme_units}}
}
\concept{internal functions that flag issues to be fixed}
