% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_complement.R
\name{reduce_complement}
\alias{reduce_complement}
\title{Reduce a list of dataframes row-binding each dataframe in a given order.}
\usage{
reduce_complement(.x, order = NULL)
}
\arguments{
\item{.x}{List of dataframes. Each should have a \code{rowid} column giving the
index of each row. Otherwise, \code{rowid} will be added with a warning.}

\item{order}{String giving the name or index of the list elements in the
order they should be row-bind, with elements on the left winning over
elements on the right.}
}
\value{
A dataframe.
}
\description{
This function orders a list of dataframes then reduces the list in order, by
row-binding each dataframe with the following one and using
\code{\link[=complement]{complement()}}.
}
\examples{
library(dplyr)

prio <- list(
  prio1 = tibble(rowid = 1:1, x = "prio1"),
  prio2 = tibble(rowid = 1:2, x = "prio2"),
  prio3 = tibble(rowid = 1:3, x = "prio3")
)
reduce_complement(prio)
# Same
reduce_complement(prio, c(1, 2, 3))

# 3 overwrites all other
reduce_complement(prio, c("prio3", "prio2", "prio1"))

# 2 overwrites over 1
reduce_complement(prio, c(2, 1))

# Adds `rowid` with a warning
prio <- list(
  prio1 = tibble(rowid = 1, x = "prio1"),
  prio2 = tibble(x = "prio2"),
  prio3 = tibble(x = "prio3")
)
reduce_complement(prio)
}
\seealso{
Other helpers: \code{\link{complement}}
}
\concept{helpers}
