% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scbi}
\alias{scbi}
\alias{scbi_tree1}
\alias{scbi_species}
\alias{scbi_stem_tiny_shrub}
\alias{scbi_stem_tiny_tree}
\title{Data from SCBI.}
\format{Objects of class tbl_df (inherit from tbl, data.frame). See structure
in \strong{Examples}.}
\source{
http://bit.ly/scbi-forestgeo-data
}
\usage{
scbi_tree1

scbi_species

scbi_stem_tiny_shrub

scbi_stem_tiny_tree
}
\description{
\itemize{
\item \code{scbi_tree1} and \code{scbi_species} are complete tree and species tables from
SCBI.
\item \code{scbi_stem_tiny_shrub} and \code{scbi_stem_tiny_tree} are small example datasets
containing exclusively shrubs and trees.
}
}
\examples{
str(scbi_tree1)

str(scbi_stem_tiny_shrub)

str(scbi_stem_tiny_tree)

str(scbi_species)
}
\concept{census datasets}
\keyword{datasets}
