% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scbi}
\alias{scbi}
\alias{scbi_tree1}
\alias{scbi_species}
\title{Data from SCBI.}
\format{Objects of class tbl_df (inherit from tbl, data.frame). See structure
in \strong{Examples}.}
\source{
http://bit.ly/scbi-forestgeo-data
}
\usage{
scbi_tree1

scbi_species
}
\description{
Data from SCBI.
}
\examples{
str(scbi_tree1)

str(scbi_species)
}
\concept{census datasets}
\keyword{datasets}
